+++ test/dummy/config/initializers/reliable_turbo_streams.rb
@@ -4,4 +4,7 @@ Rails.application.config.to_prepare do
   Turbo::StreamsChannel.include(
     ActionCable::SubscriptionAdapter::EnhancedPostgresql::ReliableBroadcasting::Channel
   )
+  Turbo::StreamsChannel.include(
+    ActionCable::SubscriptionAdapter::EnhancedPostgresql::Presence::Channel
+  )
 end